Split up the RequestALetterController into separate controllers for sending and resending
This changes reduces some confusion in the IdV flow. It will need to be made in 3 parts:

1. Adding the new routes with functionality
2. Directing to the new routes
3. Tearing down the old routes

This change is here to make sure it all works together while I do the above in separate commits

[skip changelog]
