+++ app/controllers/idv/by_mail/request_letter_controller.rb
@@ -5,38 +5,24 @@ module Idv
     class RequestLetterController < ApplicationController
       include Idv::AvailabilityConcern
       include IdvStepConcern
-      skip_before_action :confirm_no_pending_gpo_profile
       include Idv::StepIndicatorConcern
 
       before_action :confirm_mail_not_rate_limited
       before_action :confirm_step_allowed
-      before_action :confirm_profile_not_too_old
 
       def index
         @applicant = idv_session.applicant
-        @presenter = RequestLetterPresenter.new(current_user, url_options)
 
         Funnel::DocAuth::RegisterStep.new(current_user.id, current_sp&.issuer).
           call(:usps_address, :view, true)
-        analytics.idv_request_letter_visited(
-          letter_already_sent: @presenter.resend_requested?,
-        )
+        analytics.idv_request_letter_visited
       end
 
       def create
         clear_future_steps!
         update_tracking
         idv_session.address_verification_mechanism = :gpo
-
-        if resend_requested? && pii_locked?
-          redirect_to capture_password_url
-        elsif resend_requested?
-          resend_letter
-          flash[:success] = t('idv.messages.gpo.another_letter_on_the_way')
-          redirect_to idv_letter_enqueued_url
-        else
-          redirect_to idv_enter_password_url
-        end
+        redirect_to idv_enter_password_url
       end
 
       def gpo_mail_service
@@ -58,19 +44,14 @@ module Idv
 
       private
 
-      def confirm_profile_not_too_old
-        redirect_to idv_path if gpo_mail_service.profile_too_old?
-      end
-
       def update_tracking
         Funnel::DocAuth::RegisterStep.new(current_user.id, current_sp&.issuer).
           call(:usps_letter_sent, :update, true)
 
         analytics.idv_gpo_address_letter_requested(
-          resend: resend_requested?,
-          first_letter_requested_at: first_letter_requested_at,
-          hours_since_first_letter:
-            hours_since_first_letter(first_letter_requested_at),
+          resend: false,
+          first_letter_requested_at: nil,
+          hours_since_first_letter: 0,
           phone_step_attempts: RateLimiter.new(
             user: current_user,
             rate_limit_type: :proof_address,
@@ -82,63 +63,8 @@ module Idv
         ProofingComponent.find_or_create_by(user: current_user).update(address_check: 'gpo_letter')
       end
 
-      def resend_requested?
-        current_user.gpo_verification_pending_profile?
-      end
-
-      def first_letter_requested_at
-        current_user.gpo_verification_pending_profile&.gpo_verification_pending_at
-      end
-
-      def hours_since_first_letter(first_letter_requested_at)
-        first_letter_requested_at ?
-          (Time.zone.now - first_letter_requested_at).to_i.seconds.in_hours.to_i : 0
-      end
-
       def confirm_mail_not_rate_limited
         redirect_to idv_enter_password_url if gpo_mail_service.rate_limited?
-      end
-
-      def resend_letter
-        analytics.idv_gpo_address_letter_enqueued(
-          enqueued_at: Time.zone.now,
-          resend: true,
-          first_letter_requested_at: first_letter_requested_at,
-          hours_since_first_letter:
-            hours_since_first_letter(first_letter_requested_at),
-          phone_step_attempts: RateLimiter.new(
-            user: current_user,
-            rate_limit_type: :proof_address,
-          ).attempts,
-          **ab_test_analytics_buckets,
-        )
-        confirmation_maker = confirmation_maker_perform
-        send_reminder
-        return unless FeatureManagement.reveal_gpo_code?
-        session[:last_gpo_confirmation_code] = confirmation_maker.otp
-      end
-
-      def confirmation_maker_perform
-        confirmation_maker = GpoConfirmationMaker.new(
-          pii: pii,
-          service_provider: current_sp,
-          profile: current_user.pending_profile,
-        )
-        confirmation_maker.perform
-        confirmation_maker
-      end
-
-      def pii
-        Pii::Cacher.new(current_user, user_session).
-          fetch(current_user.gpo_verification_pending_profile.id)
-      end
-
-      def send_reminder
-        current_user.send_email_to_all_addresses(:verify_by_mail_letter_requested)
-      end
-
-      def pii_locked?
-        !Pii::Cacher.new(current_user, user_session).exists_in_session?
       end
 
       def step_indicator_steps
