+++ app/views/idv/by_mail/enter_code/_enter_code.html.erb
@@ -44,7 +44,7 @@
 </p>
 
 <% if @can_request_another_letter %>
-  <%= link_to t('idv.messages.gpo.resend'), idv_request_letter_path, class: 'display-block margin-top-4' %>
+  <%= link_to t('idv.messages.gpo.resend'), idv_resend_letter_path, class: 'display-block margin-top-4' %>
 <% end %>
 
 <hr class="margin-y-4" />
