+++ app/views/idv/by_mail/resend_letter/new.html.erb
@@ -1,0 +1,35 @@
+<% self.title = t('titles.idv.get_letter') %>
+
+<% content_for(:pre_flash_content) do %>
+  <%= render StepIndicatorComponent.new(
+        steps: step_indicator_steps,
+        current_step: :verify_address,
+        locale_scope: 'idv',
+        class: 'margin-x-neg-2 margin-top-neg-4 tablet:margin-x-neg-6 tablet:margin-top-neg-4',
+      ) %>
+<% end %>
+
+<%= render PageHeadingComponent.new.with_content(t('idv.gpo.request_another_letter.title')) %>
+<p>
+  <%= t('idv.gpo.request_another_letter.instructions_html') %>
+</p>
+<p>
+  <%= new_tab_link_to(
+        t('idv.gpo.request_another_letter.learn_more_link'),
+        help_center_redirect_url(
+          category: 'verify-your-identity',
+          article: 'verify-your-address-by-mail',
+          flow: :idv,
+          step: :gpo_send_letter,
+        ),
+      ) %>
+</p>
+
+<div class="margin-y-5">
+  <%= button_to t('idv.gpo.request_another_letter.button'),
+                idv_resend_letter_path,
+                method: 'put',
+                class: 'usa-button usa-button--big usa-button--wide' %>
+</div>
+
+<%= render 'idv/shared/back', fallback_path: idv_verify_by_mail_enter_code_url %>
