+++ spec/views/idv/by_mail/request_letter/index.html.erb_spec.rb
@@ -1,14 +1,10 @@
 require 'rails_helper'
 
 RSpec.describe 'idv/by_mail/request_letter/index.html.erb' do
-  let(:resend_requested) { false }
-  let(:user_needs_address_otp_verification) { false }
+  let(:user) { build(:user) }
   let(:go_back_path) { nil }
   let(:step_indicator_steps) { Idv::StepIndicatorConcern::STEP_INDICATOR_STEPS }
-  let(:presenter) do
-    user = build_stubbed(:user, :fully_registered)
-    Idv::ByMail::RequestLetterPresenter.new(user, {})
-  end
+  let(:idv_by_mail_only) { false }
 
   let(:address1) { 'applicant address 1' }
   let(:address2) { nil }
@@ -17,14 +13,11 @@ RSpec.describe 'idv/by_mail/request_letter/index.html.erb' do
   let(:zipcode) { 'applicant zipcode' }
 
   before do
+    allow(view).to receive(:current_user).and_return(user)
     allow(view).to receive(:go_back_path).and_return(go_back_path)
     allow(view).to receive(:step_indicator_steps).and_return(step_indicator_steps)
+    allow(FeatureManagement).to receive(:idv_by_mail_only?).and_return(idv_by_mail_only)
 
-    allow(presenter).to receive(:resend_requested?).and_return(resend_requested)
-    allow(presenter).to receive(:user_needs_address_otp_verification?).
-      and_return(user_needs_address_otp_verification)
-
-    @presenter = presenter
     @applicant = {
       address1: 'applicant address 1',
       city: 'applicant city',
@@ -67,44 +60,11 @@ RSpec.describe 'idv/by_mail/request_letter/index.html.erb' do
     end
   end
 
-  context 'letter already sent' do
-    let(:resend_requested) { true }
+  context 'idv_by_mail_only is enabled' do
+    let(:idv_by_mail_only) { true }
 
-    it 'has the right title' do
-      expect(rendered).to have_css('h1', text: t('idv.gpo.request_another_letter.title'))
-    end
-
-    it 'has the right body' do
-      expect(rendered).to have_text(
-        strip_tags(t('idv.gpo.request_another_letter.instructions_html')),
-      )
-    end
-
-    it 'includes link to help' do
-      expect(rendered).to have_link(
-        t('idv.gpo.request_another_letter.learn_more_link'),
-        href: help_center_redirect_url(
-          category: 'verify-your-identity',
-          article: 'verify-your-address-by-mail',
-          flow: :idv,
-          step: :gpo_send_letter,
-        ),
-      )
-    end
-
-    it 'does not include troubleshooting options' do
-      expect(rendered).not_to have_css('.troubleshooting-options')
-    end
-  end
-
-  context 'user needs address otp verification' do
-    let(:user_needs_address_otp_verification) { true }
-
-    it 'renders fallback link to return to verify path' do
-      expect(rendered).to have_link(
-        '‹ ' + t('forms.buttons.back'),
-        href: idv_verify_by_mail_enter_code_path,
-      )
+    it 'returns a cancel link' do
+      expect(rendered).to have_link(t('links.cancel'), href: idv_cancel_path(step: 'gpo'))
     end
   end
 end
